Move most of the Protobuf module to a public package.
This commit moves all of the auxiliary functions (encoding/decoding,
`clear()`, `merge()`, etc) into the `go/protomodule/` directory. They
are shimmed to work with the existing `skyProtoMessage` types, which
remain in the `internal/` dir and use Go reflection on go-protobuf v1
structs.

The `proto.to_yaml()` and `proto.from_yaml()` functions were not moved
because they are difficult to support as part of the stable API, and
are therefore deprecated. They will be removed in the v1.0 release.