Make prototext output deterministic
Both the protojson and prototext renderings intentionally randomize
the whitespace between attributes.  To make it possible to use printed
representations of these proto messages in golden tests, we want to
make it deterministic for normal daily use accept that future library
upgrades may need to also update the golden files.