debugger: create package
Add a simple interactive Starlark debugger. This is a copy of a debugger that has been in use at Stripe. It can be used in any starlark-go application, not necessarily one that uses skycfg.

The debugger exposes a single `breakpoint()` function to ordinary Starlark code, which starts up an interactive REPL-style debugging session, in which users can examine the Starlark thread stack and manipulate values. An example integration is in `debugger/example_test.go`, with the transcript of a sample debugging session located in `debugger/README.md`.

Stepping through code is not supported due to restrictions in starlark-go. See https://github.com/google/starlark-go/issues/304 for details.